modify: (ReplyBox: jsx markup ): components 분리 Heading Contents Footer
