--- wd-tumblr/src/components/ReplyBox/ReplyForm.js
@@ -1,1 +1,0 @@
-/** @format */
