+++ wd-tumblr/src/components/ReplyBox/ReplyList.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import DefaultUserProfileImg from '../../assets/images/pyramid.png';
 
 const ReplyList = ({ replyList }) => (
   <ul className="replys">
